+++ src/toast-component.ts
@@ -12,6 +12,12 @@ export class ToastComponent extends HTMLElement {
         this.remove();
     };
 
+    private handleActionButtonClick: EventListener = (e: Event) => {
+        const target = e.currentTarget as HTMLElement;
+        const index = parseInt(target.dataset.index);
+        this.settings.buttons[index].callback();
+    };
+
     private render() {
         this.dataset.uid = this.settings.uid;
 
@@ -41,10 +47,34 @@ export class ToastComponent extends HTMLElement {
         copyWrapper.appendChild(message);
         this.appendChild(copyWrapper);
 
+        if (this.settings.buttons.length) {
+            const actionsWrapper = document.createElement("toast-actions");
+
+            for (let i = 0; i < this.settings.buttons.length; i++) {
+                const button = document.createElement("button");
+                button.innerText = this.settings.buttons[i].label;
+                button.dataset.index = `${i}`;
+
+                for (let k = 0; k < this.settings.buttons[i].classes.length; k++) {
+                    button.classList.add(this.settings.buttons[i].classes[k]);
+                }
+
+                if (this.settings.buttons[i].ariaLabel) {
+                    button.setAttribute("aria-label", this.settings.buttons[i].ariaLabel);
+                }
+
+                button.addEventListener("click", this.handleActionButtonClick);
+
+                actionsWrapper.appendChild(button);
+            }
+
+            copyWrapper.appendChild(actionsWrapper);
+        }
+
         if (this.settings.closeable) {
             const closeButton = document.createElement("button");
             closeButton.setAttribute("aria-label", "close notification");
-            closeButton.className = "js-toast-close";
+            closeButton.className = "close js-toast-close";
             closeButton.innerHTML =
                 '<svg aria-hidden="true" focusable="false" data-prefix="far" data-icon="times" class="svg-inline--fa fa-times fa-w-10" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512"><path fill="currentColor" d="M207.6 256l107.72-107.72c6.23-6.23 6.23-16.34 0-22.58l-25.03-25.03c-6.23-6.23-16.34-6.23-22.58 0L160 208.4 52.28 100.68c-6.23-6.23-16.34-6.23-22.58 0L4.68 125.7c-6.23 6.23-6.23 16.34 0 22.58L112.4 256 4.68 363.72c-6.23 6.23-6.23 16.34 0 22.58l25.03 25.03c6.23 6.23 16.34 6.23 22.58 0L160 303.6l107.72 107.72c6.23 6.23 16.34 6.23 22.58 0l25.03-25.03c6.23-6.23 6.23-16.34 0-22.58L207.6 256z"></path></svg>';
             closeButton.addEventListener("click", this.handleCloseClickEvent);
@@ -57,6 +87,17 @@ export class ToastComponent extends HTMLElement {
             const closeButton:HTMLButtonElement = this.querySelector(".js-toast-close");
             if (closeButton){
                 closeButton.focus();
+            }
+        }
+        if (this.settings.buttons.length) {
+            for (let i = 0; i < this.settings.buttons.length; i++) {
+                if (this.settings.buttons[i].autofocus){
+                    const button:HTMLButtonElement = this.querySelector(`button[data-index="${i}"]`);
+                    if (button){
+                        button.focus();
+                        break;
+                    }
+                }
             }
         }
     }